+++ packages/containers/.storybook/preview.js
@@ -1,12 +1,7 @@
-import React from 'react';
-import cmf from '@talend/react-cmf';
-import { ThemeProvider } from '@talend/design-system';
 import { namespaces as componentsNamespaces } from '@talend/locales-tui-components/namespaces';
 import { namespaces as formsNamespaces } from '@talend/locales-tui-forms/namespaces';
 import { namespaces as containersNamespaces } from '@talend/locales-tui-containers/namespaces';
 import { namespaces as dsNamespaces } from '@talend/locales-design-system/namespaces';
-
-export const namespaces = [componentsNamespaces, containersNamespaces, formsNamespaces];
 
 import cmfModule, { settings } from './cmfModule';
 
@@ -27,28 +22,9 @@ export const i18n = {
 	},
 };
 
-export const loaders = [
-	async () => {
-		const { store, saga, App } = await cmf.bootstrap({
-			render: false,
-			modules: [cmfModule],
-		});
-		saga.run();
-		store.dispatch(cmf.actions.settings.receiveSettings(settings));
-		return { store, App };
-	},
-];
-
-export const decorators = [
-	(Story, { loaded: { store, App } }) => (
-		<App store={store} registry={cmf.registry.getRegistry()}>
-			<ThemeProvider>
-				<React.Suspense fallback={null}>
-					<Story />
-				</React.Suspense>
-			</ThemeProvider>
-		</App>
-	),
-];
+export const cmf = {
+	modules: [cmfModule],
+	settings,
+};
 
 export const parameters = { layout: 'fullscreen' };