+++ packages/design-system/src/components/ButtonAsLink/index.ts
@@ -1,0 +1,13 @@
+import ButtonPrimaryAsLink from './variations/ButtonPrimaryAsLink';
+import ButtonSecondaryAsLink from './variations/ButtonSecondaryAsLink';
+import ButtonTertiaryAsLink from './variations/ButtonTertiaryAsLink';
+import ButtonDestructiveAsLink from './variations/ButtonDestructiveAsLink';
+import ButtonAsLink from './ButtonAsLink';
+
+export {
+	ButtonAsLink,
+	ButtonPrimaryAsLink,
+	ButtonSecondaryAsLink,
+	ButtonTertiaryAsLink,
+	ButtonDestructiveAsLink,
+};