+++ packages/design-system/src/components/Form/Field/Input/Input.Copy.tsx
@@ -1,10 +1,10 @@
 import React from 'react';
 import { useCopyToClipboard } from 'react-use';
 import { useTranslation } from 'react-i18next';
-import Button from '../../../Button';
 import FieldGroup from '../../FieldGroup';
 import Text from './Input.Text';
 import { InputProps } from './Input';
+import { AffixButton, AffixReadOnly } from './Affix';
 
 const InputCopy = React.forwardRef(
 	(
@@ -32,15 +32,20 @@ const InputCopy = React.forwardRef(
 				label={label}
 				prefix={prefix}
 				suffix={
-					!readOnly ? (
-						<Button.Icon
+					!readOnly || disabled ? (
+						<AffixButton
 							icon="talend-files-o"
 							onClick={() => copyToClipboard(inputRef.current?.value || '')}
+							hideText
 							disabled={disabled}
 						>
 							{t('FORM_COPY_COPY_TO_CLIPBOARD', 'Copy to clipboard')}
-						</Button.Icon>
-					) : undefined
+						</AffixButton>
+					) : (
+						<AffixReadOnly icon="talend-files-o" hideText>
+							{t('FORM_COPY_COPY_TO_CLIPBOARD', 'Copy to clipboard')}
+						</AffixReadOnly>
+					)
 				}
 				readOnly={!disabled}
 				disabled={!!disabled}