+++ packages/forms/.storybook/main.js
@@ -1,9 +1,3 @@
 module.exports = {
-	stories: ['../src/**/*.stories.@(js)', '../stories-core/**/index.js'],
-	addons: [
-		'@storybook/addon-links',
-		'@storybook/addon-essentials',
-		'@storybook/addon-toolbars',
-		'@storybook/addon-knobs',
-	],
+	addons: ['@storybook/addon-knobs'],
 };