+++ packages/forms/.storybook/preview.js
@@ -1,13 +1,22 @@
 import React from 'react';
-import { IconsProvider, ThemeProvider } from '@talend/design-system';
-import init, { withI18Next } from '../../../.storybook/i18n';
-import { withA11y } from '@storybook/addon-a11y';
+
+import { namespaces as dsNamespaces } from '@talend/locales-design-system/namespaces';
+import { namespaces as componentsNamespaces } from '@talend/locales-tui-components/namespaces';
+import { namespaces as formsNamespaces } from '@talend/locales-tui-forms/namespaces';
+
+export const i18n = {
+	namespaces: [...dsNamespaces, ...componentsNamespaces, ...formsNamespaces],
+	remoteLocalesMap: {
+		'design-system': 'https://unpkg.com/@talend/locales-design-system/locales/{{lng}}/{{ns}}.json',
+		'tui-components':
+			'https://unpkg.com/@talend/locales-tui-components/locales/{{lng}}/{{ns}}.json',
+		'tui-forms': 'https://unpkg.com/@talend/locales-tui-forms/locales/{{lng}}/{{ns}}.json',
+	},
+};
 
 export const parameters = {
 	actions: { argTypesRegex: '^on[A-Z].*' },
 };
-
-init({});
 
 const withFormLayout = (story, options) => {
 	if (options.kind.includes('Layout')) {
@@ -24,19 +33,4 @@ const withFormLayout = (story, options) => {
 		</div>
 	);
 };
-
-export const decorators = [
-	withA11y,
-	withFormLayout,
-	withI18Next,
-	storyFn => {
-		return (
-			<ThemeProvider>
-				<IconsProvider bundles={['https://unpkg.com/@talend/icons/dist/svg-bundle/all.svg']} />
-				{storyFn()}
-			</ThemeProvider>
-		);
-	},
-];
-
-export { globalTypes } from '../../../.storybook/i18n';
+export const decorators = [withFormLayout];